// src/domain/external/aws/enum/config-data-id.enum.ts
export enum ConfigDataIdEnum {
  ADDRESS_RESTRICTION = 'cms_address_restriction.json',
  LASTMILE_PROVIDER_IDENTIFICAITON = 'cms_lastmile_provider_identification.json',
  CHANNEL = 'cms_channel.json',
  ACCOUNT = 'cms_account.json',
  TRANSPORTER_PRODUCT = 'cms_transporter_product.json',
  LABEL_FORMAT = 'cms_label_format.json',
  TRANSPORTER = 'cms_transporter.json',
  PLATFORM = 'cms_platform.json',
  SENDER_ADDRESS = 'cms_sender_address.json',
}